+++ honoka-json/honoka-json-gson/src/main/java/de/honoka/sdk/json/gson/GsonJsonObject.java
@@ -76,7 +76,7 @@ class GsonJsonObject extends JsonObject {
 
     @Override
     public String toPrettyString() {
-        Gson gson = Common.gsonBuilder.setPrettyPrinting().create();
+        Gson gson = Common.copyBuilder().setPrettyPrinting().create();
         return gson.toJson(originalJsonObject);
     }